+++ packages/element/src/index.ts
@@ -2,5 +2,8 @@ export * from '@availity/mui-alert';
 export * from '@availity/mui-badge';
 export * from '@availity/mui-button';
 export * from '@availity/mui-divider';
+export * from '@availity/mui-icon';
 export * from '@availity/mui-link';
+export * from '@availity/mui-paper';
 export * from '@availity/theme-provider';
+export * from '@availity/mui-tooltip';