+++ packages/paper/src/lib/Paper.tsx
@@ -1,0 +1,8 @@
+import { forwardRef } from 'react';
+import { Paper as MuiPaper, PaperProps as MuiPaperProps } from '@mui/material';
+
+export type PaperProps = Omit<MuiPaperProps, 'classes' | 'style' | 'className' | 'sx'>;
+
+export const Paper = forwardRef<HTMLDivElement, PaperProps>((props, ref) => {
+  return <MuiPaper {...props} ref={ref} />;
+});